backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg[binary]>=3.2.12
python-jose[cryptography]>=3.3.0
bcrypt>=5.0.0
python-multipart>=0.0.12
pydantic>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
email-validator>=2.1.0
pyotp>=2.9.0
qrcode>=7.4.2
pillow>=10.0.0
